producer/requirements.txt
# requirements.txt
psycopg2-binary
confluent_kafka
python-dotenv
python-snappy
asyncio
websockets
requests
pydantic
shapely